Update for roster and delete condition
The staff who is assigned to job could not be deleted,
Add a calendar to each staff for roster.

+++ app/Http/Controllers/Auth/AdminStaffController.php
@@ -115,39 +115,36 @@ class AdminStaffController extends Controller
 
     public function postDelete($id) {
         $staff = Staff::query()->find($id);
-        $staff->delete();
+        $record = Staff_Assignment::where('staff_id', '=', $id)->get();
+        if($record != null) {
+            return redirect()->back()->withErrors('This staff is currently assigned to a assignment. Operation failed.');
+        }else{
+            $staff->delete();
+        }
+
         return redirect()->route('staff.index');
     }
 
     /*Calendar*/
-    public function getCalendar() {
-        $events = Booking::all();
+    public function getCalendar($id) {
+        $events = Staff_Assignment::where('staff_id', '=', $id)->get();
         $event = [];
-        /*$event[] = \Calendar::event(
-            'title',
-            true,
-            '2020-06-02',
-            '2020-06-02',
-            1,
-            [
-                'color' => 'blue'
-                ]
-    );*/
         foreach ($events as $row) {
             $enddate = $row->date."24:00:00";
             $event[] = \Calendar::event(
-                $row->booking_type->description,
+                $row->booking->booking_type->description,
                 true,
-                new \DateTime($row->date),
-                new \DateTime($row->date),
-                $row->id,
+                new \DateTime($row->booking->date),
+                new \DateTime($row->booking->date),
+                $row->staff_id,
                 [
-                    'color' => 'red'
+                    'color' => '#3E76E5',
                 ]
+
             );
         }
         $calendar = \Calendar::addEvents($event);
-        return view('Administration.staff.calendar', compact('events', 'calendar'));
+        return view('Administration.staff.roster', compact('events', 'calendar'));
     }
